feat: cache dining preferences

+++ backend/dining/views.py
@@ -4,10 +4,12 @@ from django.db.models import Count
 from django.shortcuts import get_object_or_404
 from django.utils import timezone
 from django.utils.timezone import make_aware
+from backend.utils.cache import MONTH_IN_SECONDS_APPROX
 from rest_framework import generics
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from django.core.cache import cache
 
 from dining.api_wrapper import APIError, DiningAPIWrapper
 from dining.models import DiningMenu, Venue
@@ -57,30 +59,38 @@ class Preferences(APIView):
     """
 
     permission_classes = [IsAuthenticated]
+    key_prefix = "dining_preferences"
 
     def get(self, request):
-
         preferences = request.user.profile.dining_preferences
-
-        # aggregates venues and puts it in form {"venue_id": x, "count": x}
-        return Response(
-            {"preferences": preferences.values("venue_id").annotate(count=Count("venue_id"))}
-        )
+        key = f"{self.key_prefix}:{request.user.id}"
+        cached_preferences = cache.get(key)
+        if cached_preferences is None:
+            cached_preferences = (
+                preferences.all().values("venue_id").annotate(count=Count("venue_id"))
+            )
+            cache.set(key, cached_preferences, MONTH_IN_SECONDS_APPROX)
+        return Response({"preferences": cached_preferences})
 
     def post(self, request):
-
         profile = request.user.profile
-
         preferences = profile.dining_preferences
+
+        if "venues" not in request.data:
+            return Response(
+                {"success": False, "error": "No venues provided"}, status=400
+            )
+
+        venues = [
+            get_object_or_404(Venue, venue_id=int(venue_id))
+            for venue_id in request.data["venues"]
+        ]
 
         # clears all previous preferences associated with the profile
         preferences.clear()
+        preferences.add(*venues)
 
-        venue_ids = request.data["venues"]
-
-        for venue_id in venue_ids:
-            venue = get_object_or_404(Venue, venue_id=int(venue_id))
-            # adds all of the preferences given by the request
-            preferences.add(venue)
-
+        # clear cache
+        cache.delete(f"{self.key_prefix}:{request.user.id}")
+        
         return Response({"success": True, "error": None})
